Problem: (CRO-271) disorganised ClientRPC modules (#268)
Solution: Extract RPC calls into different module

- multisig
- sync
- staking
- wallet
