+++ client-rpc/src/main.rs
@@ -1,4 +1,4 @@
-mod client_rpc;
+mod rpc;
 mod server;
 
 use structopt::StructOpt;
