+++ client-rpc/src/rpc.rs
@@ -1,0 +1,4 @@
+pub mod multisig_rpc;
+pub mod staking_rpc;
+pub mod sync_rpc;
+pub mod wallet_rpc;
